Fix Lua error messages
Before if Lua errored the message was blank in most cases (in my testing
only compile time syntax errors showed properly, likely because only the
error was pushed to the stack since it was at compile). Changing to -1
will grab the top most item on the stack which will be the error.